Made build with latest swirlds changes

+++ benchmarks/src/jmh/java/jasperdb/VirtualDataSourceBench.java
@@ -67,9 +67,9 @@ public class VirtualDataSourceBench {
         // create data source
         VirtualLeafRecordSerializer<ContractKey, ContractValue> virtualLeafRecordSerializer =
                 new VirtualLeafRecordSerializer<>(
-                        1, DigestType.SHA_384,
-                        1, DataFileCommon.VARIABLE_DATA_SIZE,ContractKey::new,
-                        1,ContractValue.SERIALIZED_SIZE,ContractValue::new,
+                        (short) 1, DigestType.SHA_384,
+                        (short) 1, DataFileCommon.VARIABLE_DATA_SIZE,ContractKey::new,
+                        (short) 1,ContractValue.SERIALIZED_SIZE,ContractValue::new,
                         true);
         dataSource = new VirtualDataSourceJasperDB<>(
                 virtualLeafRecordSerializer,
